source mikan: download torrent instead of magnet link
see #81
