+++ .github/ISSUE_TEMPLATE.md
@@ -1,0 +1,6 @@
+Platform:
+
+python version:
+
+BGmi version:
+
